requirements.txt
streamlit==1.32.0
requests==2.31.0
python-dotenv==1.0.1
langchain-community==0.0.27
sentence-transformers==2.5.1
faiss-cpu==1.8.0
pandas==2.2.1 